+++ crates/css_ast/src/types/image.rs
@@ -24,7 +24,7 @@ mod tests {
 
 	#[test]
 	fn size_test() {
-		assert_eq!(std::mem::size_of::<Image>(), 128);
+		assert_eq!(std::mem::size_of::<Image>(), 40);
 	}
 
 	#[test]